dev-lang/gnat-gpl: move error diagnostics
Closes: https://bugs.gentoo.org/879563

+++ dev-lang/gnat-gpl/gnat-gpl-2020.ebuild
@@ -52,17 +52,6 @@ DEPEND="${RDEPEND}
 S="${WORKDIR}"/${MYP}
 PDEPEND="${PDEPEND} elibc_glibc? ( >=sys-libs/glibc-2.13 )"
 
-src_unpack() {
-	if ! use bootstrap && [[ -z "$(type ${GNATMAKE} 2>/dev/null)" ]] ; then
-		eerror "You need a gcc compiler that provides the Ada Compiler:"
-		eerror "1) use gcc-config to select the right compiler or"
-		eerror "2) set the bootstrap use flag"
-		die "ada compiler not available"
-	fi
-
-	toolchain_src_unpack
-}
-
 src_prepare() {
 	if use amd64; then
 		BTSTRP=${BTSTRP_AMD64}
@@ -83,6 +72,15 @@ src_prepare() {
 	if [[ ${gnatpath} != "." ]] ; then
 		GNATMAKE="${gnatpath}/${GNATMAKE}"
 	fi
+
+	if ! use bootstrap && [[ -z "$(type ${GNATMAKE} 2>/dev/null)" ]] ; then
+		eerror "You need a gcc compiler that provides the Ada Compiler:"
+		eerror "1) use gcc-config to select the right compiler or"
+		eerror "2) set the bootstrap use flag or"
+		eerror "3) set ADA to a working gcc ada compiler"
+		die "ada compiler not available"
+	fi
+
 	if use bootstrap; then
 		rm "${WORKDIR}"/${BTSTRP}/libexec/gcc/x86_64-pc-linux-gnu/4.7.4/ld \
 			|| die
